multiDEGGs_filter

Function to be passed to the modifyX parameter of nestcv.train or nestcv.glmnet to allow nested feature selection and augmentation via differential network analysis with multiDEGGs.

Usage

multiDEGGs_filter(y, x, keep_single_genes = FALSE, nfilter = NULL)

Arguments

y

Numeric vector or factor. Response variable (outcome), i.e. the 'metadata' named vector, as passed by nestcv.train or nestcv.glmnet.

x

Predictor variables, i.e. the assayData matrix with genes in columns and IDs in rows, as passed by nestcv.train or nestcv.glmnet.

keep_single_genes

Logical, default FALSE. If TRUE, the function will return unique individual genes along with significant pairs.

nfilter

Integer. Maximum total number of predictors to return. When keep_single_genes = TRUE, this parameter limits the combined count of unique and paired predictors (i.e., length(keep_DEGGs) + nrow(pairs) <= nfilter). Predictors are included from most to least significant: for each pair, both the pair itself and the new unique variables are included until the nfilter threshold is reached. When keep_single_genes = FALSE, nfilter only limits the number of pairs returned. If NULL, no filtering is applied and the total number of predictors will depend on how many significantly different interactions are detected by multiDEGGs in that fold.

Value

a list containing two types of predictors:

single predictors (stored in the 'keep_DEGGs' vector)

paired predictors (stored in the 'pairs' dataframe) Note that nfilter limits the maximum number of engineered features returned, however this number might be lower and will depend on how many significantly different interactions will be found in each fold by multiDEGGs. If no significantly different interactions are found the function will print a '0' and switch to t-test for that fold.

Examples

library(nestedcv)
data("synthetic_metadata")
data("synthetic_rnaseqData")

# fit a regularized linear model:
fit.glmnet <- nestcv.glmnet(
  y = as.numeric(synthetic_metadata$response),
  x =  t(synthetic_rnaseqData),
  modifyX = "multiDEGGs_filter",
  modifyX_options = list(keep_single_genes = FALSE,
                         nfilter = 20),
  modifyX_useY = TRUE,
  n_outer_folds = 5,
  n_inner_folds = 6)

summary(fit.glmnet)
#> Nested cross-validation with glmnet
#> No filter
#> Modifier:  multiDEGGs_filter 
#> Outer loop:  5-fold CV
#> Inner loop:  6-fold CV
#> 100 observations, 14 predictors
#> 
#>        alpha  lambda n.filter
#> Fold 1   0.1 0.09686        7
#> Fold 2   0.1 0.12757        7
#> Fold 3   0.1 0.15570        7
#> Fold 4   0.1 0.10091        7
#> Fold 5   0.1 0.15698        7
#> 
#> Final parameters:
#>  lambda    alpha  
#> 0.07099  0.10000  
#> 
#> Final coefficients:
#>  (Intercept) TNF:TNFRSF1A    AKT2:MTOR   IL1B:IL1R2    FASLG:FAS TGFB3:TGFBR1 
#>     1.808953    -0.188626    -0.113628     0.050012    -0.034766    -0.030783 
#> MAP2K2:MAPK3  FANCD2:FAN1 
#>    -0.020656    -0.008568 
#> 
#> Result:
#>        RMSE     R.squared   Pearson.r^2           MAE   
#>     0.47925       0.05714       0.07489       0.44620   

# fit a random forest model:
fit.rf <- nestcv.train(
  y = synthetic_metadata$response,
  x = t(synthetic_rnaseqData),
  method = "rf",
  modifyX = "multiDEGGs_filter",
  modifyX_options = list(keep_single_genes = FALSE,
                         nfilter = 20),
  modifyX_useY = TRUE,
  n_outer_folds = 5,
  n_inner_folds = 6
)
#> Fitting final model using CV on whole data
#> Duration: 3.020885 secs

fit.rf$summary
#>                Reference
#> Predicted       Non_responder Responder
#>   Non_responder            58         1
#>   Responder                 0        41
#> 
#>               AUC            Accuracy   Balanced accuracy   
#>            0.9992              0.9900              0.9881